+++ database/migrations/20201203221412_add_gifapikey_to_configuration_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddGifapikeyToConfigurationTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('configuration', function (Blueprint $table) {
+            $table->string('gifapikey')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('configuration', function (Blueprint $table) {
+            $table->dropColumn('gifapikey');
+        });
+    }
+}